Start porting to squirrel, refactor mapobject to its own file
