+++ include/gridmap.hpp
@@ -14,6 +14,7 @@
 #include "renderer.hpp"
 #include "collisionline.hpp"
 #include "proxy_sqrat.hpp"
+#include "mapobject.hpp"
 
 struct GuiContext;
 class Renderer;
@@ -22,102 +23,13 @@ class InputState;
 
 namespace Oddlib { class LvlArchive; class IBits; }
 
-
 class Animation;
-
-
-struct ObjRect
-{
-    s32 x;
-    s32 y;
-    s32 w;
-    s32 h;
-
-    static void RegisterScriptBindings(sol::state& state)
-    {
-        Sqrat::Class<ObjRect> c(Sqrat::DefaultVM::Get(), "ObjRect");
-        c.Var("x", &ObjRect::x)
-         .Var("y", &ObjRect::y)
-         .Var("w", &ObjRect::h)
-         .Var("h", &ObjRect::h)
-         .Ctor();
-        Sqrat::RootTable().Bind("ObjRect", c);
-
-        state.new_usertype<ObjRect>("ObjRect",
-            "x", &ObjRect::x,
-            "y", &ObjRect::y,
-            "w", &ObjRect::h,
-            "h", &ObjRect::h);
-    }
-};
 
 class IMap
 {
 public:
     virtual ~IMap() = default;
     virtual const CollisionLines& Lines() const = 0;
-};
-
-class MapObject
-{
-public:
-    MapObject(MapObject&&) = delete;
-    MapObject& operator = (MapObject&&) = delete;
-    MapObject(IMap& map, sol::state& luaState, ResourceLocator& locator, const ObjRect& rect);
-    MapObject(IMap& map, sol::state& luaState, ResourceLocator& locator, const std::string& scriptName);
-    void Init();
-    void GetName();
-    void Init(const ObjRect& rect, Oddlib::IStream& objData);
-    void Update(const InputState& input);
-    void Render(Renderer& rend, GuiContext& gui, int x, int y, float scale, int layer) const;
-    void ReloadScript();
-    static void RegisterScriptBindings(sol::state& state);
-
-    bool ContainsPoint(s32 x, s32 y) const;
-    const std::string& Name() const { return mName; }
-
-    // TODO: Shouldn't be part of this object
-    void SnapXToGrid();
-
-    float mXPos = 50.0f;
-    float mYPos = 100.0f;
-    s32 Id() const { return mId; }
-    void Activate(bool direction);
-    bool WallCollision(f32 dx, f32 dy) const;
-    bool CellingCollision(f32 dx, f32 dy) const;
-    std::tuple<bool, f32, f32, f32> FloorCollision() const;
-private:
-    void ScriptLoadAnimations();
-    IMap& mMap;
-
-    std::map<std::string, std::unique_ptr<Animation>> mAnims;
-    Animation* mAnim = nullptr;
-    sol::state& mLuaState;
-    sol::table mStates;
-
-    void LoadScript();
-private: // Actions
-    bool AnimationComplete() const;
-    void SetAnimation(const std::string& animation);
-    void SetAnimationFrame(s32 frame);
-    void SetAnimationAtFrame(const std::string& animation, u32 frame);
-    bool FacingLeft() const { return mFlipX; }
-    bool FacingRight() const { return !FacingLeft(); }
-    void FlipXDirection() { mFlipX = !mFlipX; }
-private: 
-    bool AnimUpdate();
-    s32 FrameCounter() const;
-    s32 NumberOfFrames() const;
-    bool IsLastFrame() const;
-    s32 FrameNumber() const;
-public:
-    bool mFlipX = false;
-private:
-    ResourceLocator& mLocator;
-    std::string mScriptName;
-    std::string mName;
-    s32 mId = 0;
-    ObjRect mRect;
 };
 
 class Level
@@ -527,7 +439,7 @@ private:
     bool mIsAo;
 
     MapObject mPlayer;
-    std::vector<std::unique_ptr<MapObject>> mObjs;
+    std::vector<std::shared_ptr<MapObject>> mObjs;
 
     enum class eStates
     {
