+++ include/proxy_sqrat.hpp
@@ -4,6 +4,7 @@
 #pragma warning(push)
 #pragma warning(disable: 4702)
 #pragma warning(disable: 4244)
+#pragma warning(disable: 4800)
 #endif
 
 #ifndef SCRAT_USE_CXX11_OPTIMIZATIONS
